+++ packages/es-dev-server/demo/compatibility/stage-3-optional-chaining.js
@@ -1,0 +1,3 @@
+const foo = {};
+
+console.log('optional chaining', foo?.bar?.buz);