+++ packages/es-dev-server/demo/compatibility/stage-3-private-class-fields.js
@@ -1,0 +1,11 @@
+class PrivateClassFields {
+  #foo = 'bar';
+
+  #bar() {
+    return this.#foo;
+  }
+
+  bar() {
+    return this.#bar();
+  }
+}