+++ packages/rollup-plugin-index-html/test/snapshots/legacy-polyfills/app.js
@@ -1,4 +1,4 @@
 // @ts-nocheck
 /* eslint-disable */
 console.log('my app');
-import('./lazy-99ef5fe0.js');
+import('./lazy-ad0cfec5.js');