+++ packages/import-maps-resolve/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0](https://github.com/open-wc/open-wc/compare/@import-maps/resolve@0.1.5...@import-maps/resolve@0.2.0) (2019-11-19)
+
+
+### Features
+
+* update to use auto compatibility of es-dev-server ([f6d085e](https://github.com/open-wc/open-wc/commit/f6d085eda5a05391d1a464b9e49222c78194b0d9))
+
+
+
+
+
 ## [0.1.5](https://github.com/open-wc/open-wc/compare/@import-maps/resolve@0.1.4...@import-maps/resolve@0.1.5) (2019-10-25)