+++ packages/webpack-index-html-plugin/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.5](https://github.com/open-wc/open-wc/compare/@open-wc/webpack-index-html-plugin@1.5.4...@open-wc/webpack-index-html-plugin@1.5.5) (2019-11-19)
+
+**Note:** Version bump only for package @open-wc/webpack-index-html-plugin
+
+
+
+
+
 ## [1.5.4](https://github.com/open-wc/open-wc/compare/@open-wc/webpack-index-html-plugin@1.5.3...@open-wc/webpack-index-html-plugin@1.5.4) (2019-11-03)